refactor: restructure OAuth2 service by extracting authorization logic
- Added AuthorizationService and TokenGrantService to handle authorization and token grant processes.
- Moved consent information retrieval logic to AuthorizationService.
- Simplified OAuth2Service by removing redundant methods and dependencies.
- Improved error handling and validation for client and redirect URI.
